# src/main/resources/openapi/users-api.yaml
openapi: 3.0.1
info:
  title: Users Microservice API
  description: API for user management in Plaza de Comidas system
  version: 1.0.0
  contact:
    name: Mateo Castro
    email: mateocastro1297@gmail.com

servers:
  - url: http://localhost:8081/api/v1/users
    description: Local server

tags:
  - name: Authentication
    description: Authentication endpoints
  - name: User
    description: User operations

paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user and return JWT token
      operationId: authenticateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../openapi/schemas/schemas-auth.yaml#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-auth.yaml#/components/schemas/JwtResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

  /admin/owner:
    post:
      summary: Create a new owner
      operationId: createOwner
      tags:
        - User
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../openapi/schemas/schemas-user.yaml#/components/schemas/OwnerRequest'
      responses:
        '201':
          description: Owner created successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-user.yaml#/components/schemas/OwnerResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - email or document already exists
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
  
  /admin/user/{userId}:
    get:
      summary: Get user by Id
      operationId: getUserById
      tags:
        - User
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user to retrieve
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-user.yaml#/components/schemas/OwnerResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - email or document already exists
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

  /owners/employee:
    post:
      summary: Create employee account
      operationId: createEmployee
      tags:
        - User
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../openapi/schemas/schemas-user.yaml#/components/schemas/EmployeeRequest'
      responses:
        '200':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-user.yaml#/components/schemas/EmployeeResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not an owner
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Email or document already exists
          content:
            application/json:
              schema:
                $ref: '../openapi/schemas/schemas-error.yaml#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Enter your JWT token in the format: Bearer <token>"